src/requirements.txt
boto3>=1.34.0
twilio>=9.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-auth>=2.20.0
